Organize scripts
