+++ scripts/3_proposal_factory.js
