+++ scripts/4_deploy_all.js
