+++ scripts/constants.js
@@ -5,10 +5,10 @@ function ratio(n) {
 }
 
 module.exports = {
-  PROPOSAL_TEMPLATES: '0x634Cf9FCE1Fcb1e82D904e396b8AB629b3C17742',
+  PROPOSAL_TEMPLATES: '0xbFA5224322B2cC7849d6CB38e827Cc439eA2e33B',
   SFC_ADDRESS: '0xA87c1a650D8aCEfcf017b3Ef480ece942E1BF02b',
   EMPTY_ADDRESS: '0x0000000000000000000000000000000000000000',
-  GOVERNANCE: '0x8949F3334624f211789f0F34D83f0C5380b4f910',
+  GOVERNANCE: '0x4e671b79AFE1a9c836207cba14a62940a00D604c',
   SFC_GOVERNABLE: '0x7a5A007d0B1d040c99A0c2Bbf22052Ec10663B51',
   ratio
 };
